Vendor containers/image and its dependencies
Add my image add-storage-transport branch to the list of dependencies
that vendor.sh fetches, and its dependencies.  This ends up adding
test-related files for some of the modules that we were already
vendoring.  Try not to merge this until containers/image has support
for containers/storage, and we can switch to consuming that.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>