Add some CRI-ish utility functions
Add an API layer that wraps the creation and removal of sandbox
containers and regular containers into functions that more closely match
where CRI-O expects to hand things off to containers/storage, and which
also moves the formatting and storage of related metadata to a location
which can be shared with others that want to integrate with it.  This
API is being revised as the implementation is being written.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>