Add truncindex to look up by short IDs
Add a truncindex to the layer, image, and container stores, so that
layers, images, and containers can also be referred to by truncated
versions of their IDs in any place where we previously accepted names.

For containers, we don't add a truncated index for referring to
containers by layer ID, because there doesn't appear to be a minimum
length in truncindex, and we don't want to mistakenly treat a truncated
but not sufficiently-unique container ID as a layer ID, which would
likely resolve to some other different container.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>