Allow storage driver to use r/o and r/w root directory
Currently this is just for use with overlay.  We want to get to the point where we could
have multiple drivers for overlay, one for local storage and potentially multiple for
network storage.

The idead would be the read/writeable content would always be set on local storage while
images could also be pulled from network storage.

Signed-off-by: Dan Walsh <dwalsh@redhat.com>