storage: add comments on the defaults
And how they are detected. A bit of these environment variable is a bit
magic and may need to change a bit, such that if the caller only
provides a driver name, but not a path, etc. that the right thing is
done.

But for now, at least document that the DefaultOptions will be used, and
there is another way to influence what driver is chosen.

Signed-off-by: Vincent Batts <vbatts@hashbangbash.com>