Add godoc notes on locking and reexec
Add notes about the Store methods that return read-only and read-write
layer, image, and container stores, and how using the values they return
bypasses locking and synchronization.

Add notes to the methods which require that the reexec package was
initialized that the reexec package needs to be initialized from main().

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>