Don't run supportsOverlay check by default
This will fail on a people using containers/storage inside of a container.
For example if you run buildah inside of a docker container on a system
with devicemapper backend.  The overlay module will not be loaded.  The
exec of "modprobe overlay" fails because there are no overlay modules inside
of the containers mount namespace.  If we ignore this check, when container/storage
eventually attempts the mount command, the kernel will be smart enough to load
the overlay module and things will succeed.  If the kernel actually does not support
overlay, the mount will fail and the user will get his answer.

If a container runtime daemon like CRI-O or Docker wants to verify at start up that
overlay file system is supported, I have made SupportsOverlay a public function and the
daemons can call this at start time.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>