storage.conf: Don't specify nodev by default
This reverts the default value of
8b1a0f8d6863cf05709af333b8997a437652ec4c
but not the ability to specify options.

Doing `nodev` on `/` breaks previously valid configurations, such as using
`mock` inside a privileged container.  While it's easy for
such containers to learn to `mount -o remount,dev /`, it's a needless
compatibility break.

Rather, e.g. podman should learn to inject it only if it's known
safe to do so (e.g. the target container doesn't have CAP_MKNOD anyways).