Need to preserve the Xattr security.capabilty when chowning.
When we chown we not only loose the setuid flags but we also loose
the file capabilties.  This means in a user namespace when we chown
ping, it looses its filecap, which means it will not longer work.

This fix will check for the security capabilies and retain them if they
exist when chowning.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>