pkg/archive: truncate ModTime when we'd otherwise round it
When writing a header with no format explicitly specified, the first
thing archive/tar 1.10 does is round the header's ModTime field,
possibly up, which confuses our change-detection logic when we later go
to check if the result of untarring the archive matches the source
content.  Truncate the timestamp before that can happen.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>