lockfile_unix: add a couple of additional safety checks
Panic in Touch() if we're not holding a write lock in one of our
threads.  Hopefully that's the one we're being called from, but we
don't guarantee that.

Panic in Modified() if we're not holding some kind of lock.

As with Locked(), this doesn't keep us from getting a false positive
when it's a different thread that holds the lock, but it's better than
nothing.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>