loadMounts(): reset counts before merging just-loaded data
Before applying mount counts that we've just loaded, reset all of the
counts.

If a layer that we thought was mounted was unmounted by another process,
there won't be a record of it in the mounts list any more, so we
wouldn't reset the mount count on our record for that layer.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>