Cleanup handling of default driver name
If the user specifies drivername="" in the storage.conf
then we default to the first driver name based on priority list, usually
overla.

The problem with this, is now the storage.conf file has options that are
driver specific.  The code does not catch this so ends up looking for a driver
name of "", and finds no matching options.

This code will set the drivername that the storage layer would choose earlies,
so that storage will get the correct options out of the config file.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>