overlay: allow storing images with more than 127 layers
Note that such images can now be stored but they cannot be
pushed yet.  Supporting both use cases would be ideal for
source-container images that are known to have a large number
of layers, likely exceeding the current limit of 127 layers.

This limit is inherited from Docker to proactively restrict the
amount and size of arguments passed when mounting.  All (lower)
layers must be specified at mount-time but we only have one page
to pass them.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>