Add better documentation for the mount_program in overlay driver
Currently users get an error message saying overlay is not supported on specific
drivers, but their is no information about the mount_program in the error
messages.  This PR adds a mention of this so that users can investigate the use
of the mount_program.

Also add information to the containers-storage.conf man page to document when
the mount_program is required to be used for an overlay driver.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>