add pkg/homedir
pkg/homedir is a fork from docker/docker and part of un-dockering the
direct dependencies of the containers projects.  There is nothing wrong
at all with the Docker packages but there is an increasing risk of
non-deterministic behavior as Docker does not provide a go module and
it's release related only to the binary but not to the go libraries.
Enforcing control over that will prevent us from regressions in the
future.

c/storage seems like the best place for this package since it's at the
very end of the containers/ dependency chain.

Note that all forked files have a copyright note to correctly attribute
the original authors.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>