Always mount the layer via overlay.
This is the way it works when you have a mount-program (fuse-overlay).
No reason not to also do this when using regular overlay.

Currenty overlay is mounting in the case where there are no lowers
on a read/only mount iff their is no mount_program. i

This situation might make sense for containers but in the case of images
the image should still be mounted. Having an empty layer makes little sense
if it is read/only.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>