Allow users to override imagestores
We have a request from a user who wants to use imagestores and
to be able to update them from podman.

Currently if I setup /var/lib/shared as an additional store
and I attempt update the share with a command like:

Error: lock "/var/lib/shared/overlay-images/images.lock" is not a read-only lock

I can not because Podman is still using the additonal store.

User expected something like:
Error: overlay: Unknown option additionalimagestore

To work.

I looked further and figured overlay.imagestore would work, but it still failed.

This PR allows users to specify additionalimagestores on the command line, and to use
more likely options to make it work.

I also cleanup some of the option parsing in overlay.go to handle overlay, overlay2, and "."
prefixes.

I checked and none of the other drivers use additionalimagestores options.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>