pkg/mount: rm path normalization from Mounted
Since moby/sys/mountinfo v0.2.0, the path does not have to be
normalized, so drop our wrapper.

Note this slightly change the call semantics: previously, a non-existing
path resulting in an error (wrapped ENOENT), while now it will return
"not mounted". The change is intentional -- the call is not to check if
the path exists, it is to check whether it is mounted.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>