Log message when graphdriver is not set
Users expect the storage to handle the fact that the
driver field is not set, but this can lead to unexpected behaviour.

Caused me to spend several hours investigating an issue, where
the overlay options were not being used, because the user created
a storage.conf file without a graphdriver.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>