Let STORAGE_DRIVER env be used even after getting from configfile
Currently when we have a storage driver set in storage.conf and we have a
STORAGE_DRIVER environement variable, the value from the storage.conf file is
used. While it would be expected to use the value from the environement variable
instead of the storage.conf (which make it easy to override in kubernetes env).

This goes in line of how we do it for STORAGE_OPT later in the same function
where the env variable here overrides the value form the file.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>