Add Dev, Nlink, and Ino to pkg/system/Stat struct
podman needs to look at the Device, Inode and number of
links of a file, but needs this to work on Windows and
Darwin.  We are using this information to determine if
a file is a hard link whe tarring it up.

Add new function FileInfoToStatT to transate a FileInfor into
the StatT struct.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>