ApplyDiff: compress saved headers without concurrency
When we're applying a diff, we compress the headers and stash them
elsewhere so that we can use them to correctly reconstruct the layer if
we need to extract its contents later.

By default, the compression uses a 1MB block, and up to GOMAXPROCS
threads, which results in allocating GOMAXPROCS megabytes of memory up
front.  That can be much more than we need, especially if the system has
many, many cores.  Drop it down to 1 megabyte.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>