Separate the IDMappingOptions logic from other LayerOptions work
For some reason, store.PutLayer does not pass _all_ of LayerOptions
to layerStore.Put; preserve that, but modify the code so that we
can pass more values in the future, without duplicating the logic.

Also deal with the local struct directly instead indirecting through
a pointer. (The compiler can probably optimize that to the same thing,
but having a non-pointer variable at least eliminates the analysis
to see if it can be nil.)

Should not change behavior.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>