Use math/rand.Rand.Read instead of unsafe slice manipulation
The slice manipulation _is_ actually almost 50% faster,
but the implementation is only correct for sizes that are
a multiple of 8, the use of unsafe is somewhat scary, and
there's only exactly one caller, with the total difference
in the order of tens of milliseconds.

Let's do the simple thing instead.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>