GetDataHome accept uid
GetDataHome looks for the home directory on the user executing the function, if you want to get
the rootless storage config while running as root, this function fails to return the rootless home dir.
Adding a uid and checking if that uid matches ours, allows us to see if we need to find a different user's home dir.

Signed-off-by: cdoern <cdoern@redhat.com>