pkg: idtools: export RawTo{Container,Host}
While the IDMapping methods are preferable for most users, sometimes it
is necessary to map a single ID using a given mapping. In particular
this is needed for psgo to be able to map the user and group entries in
/proc/$pid/status using the user namespace of the target process.

Required to resolve CVE-2022-1227 for Podman v1.9.3.

NOTE: this commit is going into a Podman-only branch.  Podman v1.9.3 is
using c/storage v1.19.1.  v1.19.2 has been released with unrelated
changes which are not approved for RHEL.  Hence, we need to vendor the
fixes by commit (rather than a version) into the v1.9 branch in Podman.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>
Backported-by: Valentin Rothberg <vrothberg@redhat.com>