pkg: idtools: export RawTo{Container,Host}
While the IDMapping methods are preferable for most users, sometimes it
is necessary to map a single ID using a given mapping. In particular
this is needed for psgo to be able to map the user and group entries in
/proc/$pid/status using the user namespace of the target process.

Required to resolve CVE-2022-1227 for Podman v3.2.3 in RHEL which is
using c/storage v1.31.x.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>
Signed-off-by: Valentin Rothberg <vrothberg@redhat.com>