pkg: idtools: export RawTo{Container,Host}
While the IDMapping methods are preferable for most users, sometimes it
is necessary to map a single ID using a given mapping. In particular
this is needed for psgo to be able to map the user and group entries in
/proc/$pid/status using the user namespace of the target process.

Required to resolve CVE-2022-1227 for Podman v3.4.2 in RHEL which is
using c/storage v1.36.x.  Note that Podman v3.4.X in Fedora uses
c/storage v1.37.x.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>
Backported-by: Valentin Rothberg <vrothberg@redhat.com>