Fix an ignored error return value.
The "if err == nil" is always true, because
there are no prior assignments to err; but it
creates a new scope, making the "err = r.Save()"
call write to an ignored variable.

Remove the unnecessary condition to avoid that.

(Note that the code still continues on after a failure
to save, i.e. it returns a reference to a created in-memory-only
container along with the error value.)

Signed-off-by: Miloslav Trmač <mitr@redhat.com>