image, layers, container: append new names to the ones which were written
We are deleting names which were already written in store.

This creates faulty behviour when builds are invoked in parallel manner, as
this removes names for other builds.

To fix this behviour we must append to already written names and
override if needed.

Following patch fixes concurrent builds via podman or buildah.

Signed-off-by: Aditya R <arajan@redhat.com>