Preserve extra overlayfs mount options when loading many-layered images
When overlayfs loads a container image that has many layers, each becomes
an extra argument to the 'lowerdir' list, and if it has many many many
of them it will try to squeeze more in by rewriting them as relative paths.

Unfortunately, this case was forgetting 'userxattr' and 'volatile' and
any other user-supplied mount options. It meant that in rootless mode
(which needs userxattr to behave properly) you'd get an EXDEV error
anytime you tried to rename() a directory.

Fixes https://github.com/containers/podman/issues/13123

Signed-off-by: Nick Guenther <nick.guenther@polymtl.ca>