ExecRunnable: prevent logging on stderr when its not expected
ExecRunnable is invoked by buildah and error from c/storage is logged to
stdout. This creates a drift between the between the behavior which is
generated by rootless and rootful buildah.

This corrects the behavior for rootless buildah so it has parity with
buildah invoked for root user.

This helps in https://github.com/containers/buildah/pull/3804

Signed-off-by: Aditya R <arajan@redhat.com>