chown: use ToHostOverflow when chowning
when chowning an image, fall back to the overflow ID when a UID or GID
cannot be mapped to the target user namespace.

This ensures the chown driver works similar to what we do with
idmapped mounts when it is supported for overlay.

It is needed for CRI-O to support user namespaces in Kubernetes since
the Kubelet picks a static size for the user namespace and it might
break some images using IDs outside the picked range.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>