layers.Wipe(): try to remove layers before their parents
When removing layers, try to remove layers before removing their
parents, as the lower-level driver may enforce the dependency beyond
what the higher-level logic does or knows.

Do this by sorting by creation time as an attempt at flattening the
layer tree into an ordered list.  It won't be correct if a parent layer
needed to be recreated, but it's more likely to be correct otherwise.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>