Remove a manual layerStore.Touch() call from store.Shutdown()
It's completely unclear why this is necessary; the unmount calls
actually write to the separate mountpoints.json file, with a separate lock.

When this was originally introduced in b046fb5a9a1ad6d25bd6c854400f090131c8a21c ,
there wasn't a separate mountpoints.json file, and store.go was manually
calling Touch() everywhere; so my best guess is that this is just an
artifact of making those two API changes, and the Touch() is not actually
necessary.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>