Correctly handle lockfiles without writes
A newly created lock-file is of size zero, which happens for example
if you reset the storage state. In this state, lockfile.Modified()
always returns true, because it thinks any short read means something
changed.

I ran into this with the image store, which was empty (using images
from a different store), and it keept re-reading the image store all
the time even if nothing changed (i.e. both lock file reads returned
zero bytes).

The proper way to handle this seems to be to compare partial reads
also.

Signed-off-by: Alexander Larsson <alexl@redhat.com>