overlay: drop relLowers workaround
commit 7c5964df95c892cfbdbce594cf5a8e2973c70fd7 changed the way the
mountFrom program works by always using the /proc/self/fd/%d path for
the lower dirs.  Hence there is no gain to pass a path relative to the
storage, but we can just pass the absolute path as the resulting file
path length is the same and won't affect how many layers we can
specify.

This fixes idmapped mounts when the total length of lower layers is >
4096 and it also solves the problem of using layers from additional
stores.

Closes: https://github.com/containers/storage/issues/1410

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>