Rework in-process concurrency control of layerStore
Instead of basing this on exclusivity loading via loadMut (which was incorrect,
because contrary to the original design, the r.layerspathModified
check in r.Modified() could trigger during the lifetime of a read lock)
use a very traditional read-write lock to protect the fields of imageStore.

Also explicitly document how concurrent access to fields of imageStore
is managed.

Note that for the btrfs and zfs graph drivers, Diff() can trigger
Mount() and unmount() in a way that violates the locking design.
That's not fixed in this PR.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>