Don't cross filesystem boundaries in ChangesDirs
This is needed on FreeBSD to support layer diffs for running containers.
On FreeBSD, the container and host share a mount namespace and the changes
are generated from the same filesystem mount as the running container.
Since the container will have sub-mounts for e.g. /dev and mounted volumes,
we need to exclude these from the diff.

Signed-off-by: Doug Rabson <dfr@rabson.org>