Fix concurrent accesses to store.graphDriverName
Don't modify store.graphDriverName on reloads, so that it is constant
throughout the life of the store, and it can be accessed without locking,
as it has actually been done (to this point incorrectly).

That requires special-casing the initial load(); so, split
the actual driver creation into a createGraphDriverLocked() method.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>