Fix locking in store.canUseShifting
This was using the graphDriver field without locks, and the graph driver itself,
while the implementation assumed exclusivity.

Luckily all callers are actually holding the layer store lock for writing, so
use that for exclusion. (layerStore already seems to extensively assume
that locking the layer store for writing guarantees exclusive access to the graph driver,
and because we always recreate a layer store after recreating the graph driver,
that is true in practice.)

Signed-off-by: Miloslav Trmač <mitr@redhat.com>