layer store: Don't reload the layer store after we save it
The lockfile we use propertly handles the case that we Touch() it. In
other words, a later Modified() call will return false.

However, we're also looking at the mtime, which was failing. This
uses the new AtomicWriteFileWithOpts() feature to also record the
mtime of the file we write on updates.

Signed-off-by: Alexander Larsson <alexl@redhat.com>