check: suppress certain errors when pull_options might cause them
Files hard linked in from an OSTree repository won't tend to have the
right timestamps on them, so we have to accept that they'll not match
what we have on record when we're using one to speed up pulling images.

If we're ignoring chown errors when populating layers, then there's no
point in expecting the ownerships of the contents of on-disk layers to
match expectations.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>