On Read/only additional stores, ignore Read/Only errors
We want to setup empty additonal stores which could potentially
be stored on a read-only mount point (ostree).  Currently empty
stores work fine as long as the directory is read/write, because
when we attempt to lock them, the code wants to create a lock file
directory. This returns an error on read-only directories.

This patch ignores read-only addiitonal stores with no content.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>