overlay: use `dir` for final `diff`/`merged` dirs
This commit fixes an issue with the `overlay.go:Driver.get()` code
when the requested `id` comes from an extra storage dir that got
specified with e.g. `podman --imagestore`. In this scenario the
final layer with the `.../diff` dir must come from the extra
storage dir instead of from the main storage (that is may not
contain this id).

This fixes issue #1779 for my test case and I see no other issues
in my "light" testing for the "normal" cases. However I'm not
familiar with this code base and this is (of course) an extremly
sensitive part of the code so please have a careful look.

More details in https://github.com/containers/storage/issues/1779

Signed-off-by: Michael Vogt <michael.vogt@gmail.com>