pkg/parsers/kernel: simplify
Since golang.org/x/sys/unix provides portable way to use uname(2)
syscall, let's use it for all unix variants, removing the duplicated
code from uname_*.go.

While at it:
 - remove uname_unsupported*.go (not used);
 - use unix.ByteSliceToString.

NOTE this also removes the public definition of struct Utsname. I am
puzzled as to why it was exported as it is not (and was not) used by
any public functions in this package, nor I was able to find any usage
of kernel.Utsname (using sourcegraph).

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>