layers: allow mount in ro stores
There is nothing conceptually wrong in this as the Mount itself does
not modify the store.

newFileGetter() works today with overlay only because it uses the
drivers.DiffGetterDriver shortcut that reads files directly from their
storage.

graph drivers that do not allow that, need to mount the layer instead.
composefs could be changed to use the same code as overlay, but it
requires more refactoring to propagate the TOC information down to the
overlay layer.  For now just fix mounting from a ro store, as
newFileGetter() already assumes anyway.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>