Add doc comments around "string bool" usage
I've been poking at composefs related things here and
stumbled across the very confusing fact that some values
in the storage config are booleans - but must be
*strings* even though TOML supports native boolean values.

Further, digging in more we actually use two different
parsers for boolean values; one silently accepts anything
that is not the string `"true"` to mean `false`, but
the other uses the Go strconv function and does return
an error.

I'm not trying to fix any of this yet; the clear fix
is to widen what we accept into a native TOML boolean type
*in addition* which would require a bit of custom deserialization.
Let's just document this to start.

Signed-off-by: Colin Walters <walters@verbum.org>