Ignore rootless EPERM failures setting security xattrs
This is specifically for the IMA xattrs, which cannot be set
except as root; as rootless, they fail with EPERM. We never
noticed this before (likely because IMA xattrs seem uncommon in
the wild) but if there is a file in an image with an IMA xattr
rootless Podman becomes completely unable to use the image. This
is particularly relevant because the catatonit binary Podman uses
for building its pause image has started to include an IMA xattr
on Fedora Rawhide, which is breaking rootless Podman there rather
badly.

Since this cannot work as rootless, it seems simplest to try to
set the xattr, but tolerate failure iff the error is EPERM and
we are not run as root.

Fixes: https://github.com/containers/podman/issues/18543

Signed-off-by: Matt Heon <mheon@redhat.com>