Treat XDG_CONFIG_HOME and $HOME/.config the same way
Currently we handle XDG_CONFIG_HOME and the fallback $HOME/.config
differently. We create the later if it does not exist and verify that
it is owned by the current user. This patch makes XDG_CONFIG_HOME
follow the same pattern.

Also Deprecate all GetDATA() functions and just use DATA().

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>