Update copy.DirCopy to not unlink sockets
- Closes #2113

`copy.DirCopy` contains code to create named sockets matching those on
the source; however, this operation closes the socket, removing it
implicitly from the destination and causing later chown and chmod
operations to fail.

This commit removes configures the socket to not unlink the underlying
socket file from the file system during `.Close()`, leaving the socket
in place at the destination, as it is in the source.

Signed-off-by: Jonathon Anderson <janderson@ciq.com>