Strip PROJINHERIT flag from top-level dir of quotas
Basically, PROJINHERIT causes all lower-level directories to get
the same project ID. This is a good thing for the directories
that are supposed to have quotas. It is not a nice thing for the
top-level directory. We set a project ID on that directory so we
know what the base ID is for our subdirectories to use, but we do
not want that ID propagated to subdirectories else everything
will end up using that single quota ID. Stripping the flag from
the top-level directory (if it is present) resolves this.

Partial fix for https://issues.redhat.com/browse/RHEL-18038

Full fix also requires Podman changes as we were setting quotas
on the incorrect directory.

Signed-off-by: Matt Heon <mheon@redhat.com>