Don't set UncompressedSize on chunked pull
The current value obtained by summing the sizes of regular file contents
does not match the size of the uncompressed layer tarball.

We don't have a convenient source to compute the correct size
for estargz without pulling the full layer and defeating the point.

For recent zstd:chunked images, we have the full tar-split,
so we can compute the correct size; for now, this doesn't do that.
That might slow down image size computation.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>