homedir: check runtime dir validity using system boot_id
Add a boot_id check to ensure the runtime directory is not reused after
a system reboot.  This prevents the use of stale runtime directories
when the runtime directory is created by pkh/homedir because the user
is running without a proper user session.

If the file /proc/sys/kernel/random/boot_id is not present, the check
is ignored.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>