Bump maxTocSize to 150 MB
We have seen an image with:
- total size 1.43 GB
- uncompressed zstd:chunked manifest size of 91.7 MB
- uncompressed tar-split size (not constrained by maxTocSize) 310 MB

Without more infrastructure, we are just guessing about what
the system we are running on can support, so, for now, *shrug*, bump
the number.

Eventually we should stream the data from/to disk, making this
much less relevant; that makes building the infrastructure to
estimate available memory unattractive.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>