homedir: check ConfigHome is writable by current user
This replaces the check that the user config directory is owned by the
current user with a check that the directory is writable by the current
user. Usually these go together, but in the case that they differ,
presumably the important thing is whether the current user can read from
and write to the directory, regardless of whether they own it.

Also fixed a dead link in a comment.

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>