[release-1.59] pkg/archive.tarWriter.addFile(): don't change hardlink targets
When examining a file that we're adding to an archive to see if it's
hard linked to another file that's already been written, or to one that
might be written later, only set an entry in the inode->name map that
points to the entry we're writing for the first file.  This keeps us
from writing tar headers with Typeflag==TypeLink and a Linkname that
points to another entry with Typeflag==TypeLink in cases where the inode
has more than two names.

Addresses: https://github.com/containers/buildah/issues/6297 for the
release-1.59 branch and eventualy Buildah v1.41.* and Podman v5.6.*

Cherry-pick of https://github.com/containers/storage/pull/2357

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>
Signed-off-by: tomsweeneyredhat <tsweeney@redhat.com>